fix: point develop-added files at the moved pkg/skychat/group (rebase fallout)
