+++ cmd/apps/skychat/commands/group.go
@@ -23,8 +23,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/skycoin/skywire/cmd/apps/skychat/group"
 	"github.com/skycoin/skywire/pkg/cipher"
+	"github.com/skycoin/skywire/pkg/skychat/group"
 	"github.com/skycoin/skywire/pkg/skychat/xfer"
 	"github.com/skycoin/skywire/pkg/visor"
 )
